Create tracing with dot
This mirrors the image for easy writing, and uses a dot to write. 
Things to change is the contour line for color recognition in the mask window, and getting the dot to start/stop with a gesture.
Observe the code and understand how to the writing works so we can dissect pieces into our hand tracking code.